# Test for and fix (*vm).Interrupt data race