--- a/vm.go
+++ b/vm.go
@@ -6,6 +6,7 @@ import (
 	"math"
 	"strconv"
 	"sync"
+	"sync/atomic"
 )
 
 const (
@@ -111,7 +112,7 @@ type vm struct {
 	stashAllocs int
 	halt        bool
 
-	interrupt     bool
+	interrupt     uint32
 	interruptVal  interface{}
 	interruptLock sync.Mutex
 }
@@ -277,26 +278,30 @@ func (vm *vm) init() {
 
 func (vm *vm) run() {
 	vm.halt = false
-	for !vm.halt && !vm.interrupt {
+	for !vm.halt && !vm.interrupted() {
 		vm.prg.code[vm.pc].exec(vm)
 	}
 
-	if vm.interrupt {
+	if vm.interrupted() {
 		vm.interruptLock.Lock()
 		v := &InterruptedError{
 			iface: vm.interruptVal,
 		}
-		vm.interrupt = false
 		vm.interruptVal = nil
+		atomic.StoreUint32(&vm.interrupt, 0)
 		vm.interruptLock.Unlock()
 		panic(v)
 	}
 }
 
+func (vm *vm) interrupted() bool {
+	return atomic.LoadUint32(&vm.interrupt) != 0
+}
+
 func (vm *vm) Interrupt(v interface{}) {
 	vm.interruptLock.Lock()
 	vm.interruptVal = v
-	vm.interrupt = true
+	atomic.StoreUint32(&vm.interrupt, 1)
 	vm.interruptLock.Unlock()
 }