@nilium 2017-02-15
@dop251 If I run a program with the race detector on and goja triggers a series of error messages in the output, that _should_ be enough to justify fixing it. It should also be alarming, not something to be content with (much less deliberate, which is questionable at best). That means the program is not correct because it's -- as it suggests -- a race.

If nothing else, what seebs is referring to is that anything that isn't covered by the [Go memory model](https://golang.org/ref/mem), is effectively undefined and may result in anything from incorrect behavior to reindeers to panics (this is because it's a data race -- Go can't define what happens in those scenarios). Unless you can amend the race detector and spec, goja is incorrect. And, per the doc:

> If you must read the rest of this document to understand the behavior of your program, you are being too clever.
>
> Don't be clever.

Anyhow, I ran the benchmarks included in goja to see if this would have any impact. I can see no significant difference in the results on my laptop outside of the fibonacci benchmark, which I assume is broken since it's only managed one run. The empty loop may have suffered the most, with a difference of about 300ns (or about 0.3 microseconds), which isn't significant enough for me to say the interrupt fix has "impact[ed] performance quite significantly." The master columns refer to master as it is now, the patched columns refer to the benchmark times with a correct interrupt:

```
                      |     master |        patched |              master |            patched
----------------------+------------+----------------+---------------------+-------------------
BenchmarkCompile      |        500 |            500 |       2739791 ns/op |      2752266 ns/op
BenchmarkGoReflectGet |    1000000 |         500000 |          2572 ns/op |         2570 ns/op
BenchmarkPut          |   30000000 |       30000000 |          46.2 ns/op |         49.0 ns/op
BenchmarkPutStr       |   30000000 |       30000000 |          39.9 ns/op |         41.8 ns/op
BenchmarkGet          |  100000000 |      100000000 |          16.2 ns/op |         16.2 ns/op
BenchmarkGetStr       |  100000000 |      100000000 |          12.5 ns/op |         10.9 ns/op
BenchmarkToString1    | 2000000000 |     2000000000 |          0.63 ns/op |         0.65 ns/op
BenchmarkToString2    |  500000000 |     1000000000 |          3.20 ns/op |         3.06 ns/op
BenchmarkConv         | 2000000000 |     2000000000 |          0.32 ns/op |         0.33 ns/op
BenchmarkArrayGetStr  |   50000000 |       50000000 |          23.9 ns/op |         27.8 ns/op
BenchmarkArrayGet     |   20000000 |       20000000 |          66.4 ns/op |         72.6 ns/op
BenchmarkArrayPut     |  100000000 |      100000000 |          11.3 ns/op |         11.9 ns/op
BenchmarkToUTF8String | 1000000000 |      500000000 |          2.88 ns/op |         3.26 ns/op
BenchmarkAdd          |   50000000 |       50000000 |          32.5 ns/op |         32.5 ns/op
BenchmarkAddString    |    5000000 |        5000000 |           260 ns/op |          259 ns/op
BenchmarkVmNOP2       |  200000000 |      200000000 |          6.51 ns/op |         6.65 ns/op
BenchmarkVmNOP1       |   50000000 |       50000000 |          37.9 ns/op |         37.8 ns/op
BenchmarkVmNOP        |  100000000 |      100000000 |          11.8 ns/op |         16.9 ns/op
BenchmarkVm1          |   30000000 |       30000000 |          50.3 ns/op |         56.8 ns/op
BenchmarkFib          |          1 |              1 |    6836786055 ns/op |   7250626136 ns/op
BenchmarkEmptyLoop    |     100000 |         100000 |         16830 ns/op |        17151 ns/op
BenchmarkVMAdd        |  100000000 |      100000000 |          22.6 ns/op |         22.5 ns/op
```

So, again, I urge you to prioritize fixing the data race instead of trying to rationalize it as benign. I'm happy to trade 300ns for sane code, but I would never be happy to gain 300ns at the cost of a data race. In go, there are no benign data races, and goja having one in its interrupt code is not acceptable when it should be safe to interrupt a VM.